Long to lowercase except preferences

+++ backend/src/main/java/nl/hva/backend/repositories/NotesRepository.java
@@ -19,7 +19,7 @@ public class NotesRepository {
         return namedQuery.getResultList();
     }
 
-    public Note findNoteById(int noteId) {
+    public Note findNoteById(long noteId) {
         return entityManager.find(Note.class, noteId);
     }
 
@@ -29,7 +29,7 @@ public class NotesRepository {
         return entityManager.merge(note);
     }
 
-    public boolean deleteNoteById(int noteId) {
+    public boolean deleteNoteById(long noteId) {
         Note note = findNoteById(noteId);
         if (note != null) {
             entityManager.remove(note);
